+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rehash GUI after importing messages
 - Add latest message view
 - Add anonymous/non-anonymous fetch
+- Inline image support
 
 ### Changed
 - Updated GUI toolkit to Fyne v2.0.3